day 17: compute length outside the recurrence loop
When I fed back the answer from the previous commit into the program, I
noticed it actually produced output that was one-too-long.  This was
because it was computing the Computer's output with our original value
of `a`, but we were then returning the next value of a instead.

And then I also realized that if we're not changing `a`, then we don't
need to re-do the Computer evaluation every time on the inside loop --
shrink that loop to just the recurrence term instead.