day 16: only visit each reachable node once
This fundamentally depends on the observation that, for any node X along
a shortest path, the shortest distance from S to X and from X to E must
add up to be the exact same value... any more-complicated way would
necessarily be longer.

This still took about 70s to run, but it did return the correct answer!