day 18: use DFS for the part 2 path
It actually doesn't matter *how* good the path actually is, just that it
exists.  `dfs_end` is essentially the same as `find_end`, except it now
just pushes and pops to/from the end of a Vec -- we don't need the Deque
anymore.

Now we're down from 0.2s to 0.08s because it doesn't spend time
accumulating middle nodes for potential short paths.